Throwing in the towl.
Calling it done and correct because at this point I can't honestly say
whether it is or isn't. A massive rewrite is imminent.

It looks like the Sourceposition is off on the endings of tokens. For
the end of a token position, it should be set to
currentToken.position.finish rather than currentToken.position.start